create UserContext, integrate context with User Profile components, add buttons to UserSideBar
