+++ context/UserContext.tsx
@@ -1,0 +1,28 @@
+import { User } from '@/app/profile/components/EditUserDetailsDialog';
+import React, { createContext, useContext, useState, ReactNode } from 'react';
+
+
+interface UserContextProps {
+    user: User | undefined;
+    setUser: React.Dispatch<React.SetStateAction<User | undefined>>;
+}
+
+const UserContext = createContext<UserContextProps | undefined>(undefined);
+
+export const UserProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
+    const [user, setUser] = useState<User>();
+
+    return (
+        <UserContext.Provider value={{ user, setUser }}>
+            {children}
+        </UserContext.Provider>
+    );
+};
+
+export const useUser = () => {
+    const context = useContext(UserContext);
+    if (!context) {
+        throw new Error('useUser must be used within a UserProvider');
+    }
+    return context;
+};
